# back/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Piyachok"
  },
  "paths": {
...
    "/users/{user_id}": {
      "get": {
        "summary": "Get one user",
        "tags": [
          "Users"
        ],
        "operationId": "getOneUser",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "type": "string",
            "required": true
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "description": "Set `access_token` from /auth",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/singleUser"
            },
            "description": "You got one user"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Put one user",
        "operationId": "updateOneUser",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Cоня Рибська",
                  "description": "username"
                },
                "email": {
                  "type": "string",
                  "example": "sofaribska@gmail.com",
                  "description": "user email"
                },
                "picture": {
                  "type": "string",
                  "example": "https://lh3.googleusercontent.com/a/AEdFTp6_rN4O1Ov60r9HOguESrCn8OMHvDsHhN3wpnPYxg=s96-c",
                  "description": "user google avatar"
                },
                "admin": {
                  "type": "boolean",
                  "example": true,
                  "default": "false",
                  "description": "is user admin"
                }
              }
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "type": "number",
            "required": true
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "description": "Set `access_token` from /auth",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "User is updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete one user",
        "operationId": "deleteOneUser",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "type": "number",
            "required": true
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "description": "Set `access_token` from /auth",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "User is deleted"
          }
        }
      }
    },
...
    "/reviews/{id}": {
      "get": {
        "summary": "Get reviews by establishment id",
        "tags": [
          "Reviews"
        ],
        "operationId": "getReviewsByEstablishmentId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "page",
            "in": "query",
            "type": "number"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "number"
          },
          {
            "name": "sort",
            "in": "query",
            "enum": [
              "created_at-DESC",
              "created_at-ASC"
            ],
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "reviews":{
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/singleReviewWithUser"
                  }
                },
                "count": {
                  "type": "number",
                  "example": 32
                }
              }
            },
            "description": "You got reviews by establishment id"
          }
        }
      }
    },
...
    "/reviews/users/{id}": {
      "get": {
        "summary": "Get reviews by user id",
        "operationId": "getReviewsByUserId",
        "tags": [
          "Reviews"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "description": "Set `access_token` from /auth",
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "type": "number",
            "required": true
          },
          {
            "name": "page",
            "in": "query",
            "type": "number"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "reviews": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/singleReviewWithEstablishment"
                  }
                },
                "count": {
                  "type": "number",
                  "example": 27
                }
              }
            },
            "description": "You got reviews by user id"
          }
        }
      }
    }
  },
  "definitions": {
    "singleUser": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "number",
          "example": "20",
          "description": "user id"
        },
        "name": {
          "type": "string",
          "example": "Cоня Рибська",
          "description": "username"
        },
        "email": {
          "type": "string",
          "example": "sofaribska@gmail.com",
          "description": "user email"
        },
        "picture": {
          "type": "string",
          "example": "https://lh3.googleusercontent.com/a/AEdFTp6_rN4O1Ov60r9HOguESrCn8OMHvDsHhN3wpnPYxg=s96-c",
          "description": "user google avatar"
        },
        "admin": {
          "type": "boolean",
          "example": true,
          "default": "false",
          "description": "is user admin"
        },
        "createdAt": {
          "type": "string",
          "example": "2023-01-25 17:53:48.940 +00:00",
          "description": "when user was created"
        },
        "updatedAt": {
          "type": "string",
          "example": "2023-01-25 17:53:48.940 +00:00",
          "description": "when user was updated"
        }
      }
    },
    "singleEstablishment": {
      "type": "object",
      "properties": {
        "establishment_id": {
          "type": "number",
          "example": 112
        },
        "title": {
          "type": "string",
          "example": "Prime"
        },
        "type": {
          "type": "string",
          "example": "Restoraunt"
        },
        "avatar": {
          "type": "string",
          "example": "users\\30\\328\\establishment_photos\\41bbc2b0-9dbc-11ed-b0f9-cb97cc80cc9f.jpg"
        },
        "location": {
          "type": "string",
          "example": "Lychakivska Street, 23, Lviv, Lviv Oblast, Ukraine"
        },
        "start_work": {
          "type": "string",
          "example": "19:55:00"
        },
        "end_work": {
          "type": "string",
          "example": "23:58:00"
        },
        "view_statistics": {
          "type": "number",
          "example": 10
        },
        "tags": {
          "type": "array",
          "example": [
            "bar",
            "hookan"
          ]
        },
        "average_check": {
          "type": "number",
          "example": 1000
        },
        "approved": {
          "type": "number",
          "example": false
        },
        "pending": {
          "type": "number",
          "example": false
        },
        "rejected": {
          "type": "number",
          "example": true
        },
        "user_id": {
          "type": "number",
          "example": "Lviv"
        },
        "photos": {
          "type": "array",
          "items": {
            "type": "string",
            "default": "users\\30\\328\\establishment_photos\\41bbc2b1-9dbc-11ed-b0f9-cb97cc80cc9f.jpg"
          }
        },
        "phone": {
          "type": "number",
          "example": "+380986649500"
        },
        "createdAt": {
          "type": "number",
          "example": "2023-01-25 17:53:48.940 +00:00"
        },
        "updatedAt": {
          "type": "number",
          "example": "2023-01-25 17:53:48.940 +00:00"
        }
      }
    },
...
    "singleReviewWithEstablishment": {
      "type": "object",
      "properties": {
        "review_id": {
          "type": "number",
          "example": 2
        },
        "text": {
          "type": "string",
          "example": "Good place"
        },
        "check": {
          "type": "number",
          "example": 1500
        },
        "user_id": {
          "type": "number",
          "example": 23
        },
        "establishment_id": {
          "type": "number",
          "example": 119
        },
        "createdAt": {
          "type": "string",
          "example": "2023-01-25 17:53:48.940 +00:00",
          "description": "when user was created"
        },
        "updatedAt": {
          "type": "string",
          "example": "2023-01-25 17:53:48.940 +00:00",
          "description": "when user was updated"
        },
        "establishment": {
          "$ref": "#/definitions/singleEstablishment"
        }
      }
    },
    "singleReviewWithUser": {
      "type": "object",
      "properties": {
        "review_id": {
          "type": "number",
          "example": 117
        },
        "text": {
          "type": "string",
          "example": "Good place"
        },
        "check": {
          "type": "number",
          "example": 1500
        },
        "user_id": {
          "type": "number",
          "example": 23
        },
        "establishment_id": {
          "type": "number",
          "example": 119
        },
        "createdAt": {
          "type": "string",
          "example": "2023-01-25 17:53:48.940 +00:00",
          "description": "when user was created"
        },
        "updatedAt": {
          "type": "string",
          "example": "2023-01-25 17:53:48.940 +00:00",
          "description": "when user was updated"
        },
        "user": {
          "$ref": "#/definitions/singleUser"
        }
      }
    }
  }
}